+++ features/ui/index.ts
@@ -1,4 +1,4 @@
-export * from "./abadge";
-export * from "./abutton";
+export * from "./badge";
+export * from "./button";
 export * from "./page-container";
 export * from "./sidebar-navigation";
